Moved db tasks to their own file
