+++ tasks/__init__.py
@@ -9,6 +9,7 @@ from . import i18n
 from . import servers
 from . import docs
 from . import test
+from . import db
 
 ns.add_collection(prereqs)
 ns.add_collection(i18n)
@@ -16,3 +17,4 @@ ns.add_collection(servers)
 ns.add_collection(assets)
 ns.add_collection(docs)
 ns.add_collection(test)
+ns.add_collection(db)
